Add files via upload
The accuracy graph of the models using the LSTM layer is presented in 1 shows the bar charts that can be used to infer the accuracy of each of the four models constructed. From observation, we can infer that the fourth model gives an accuracy of 99%. The first and the fourth bar chart presented, are similar as we used similar configuration of the neural network layers to train the model. Model 1 and model 4 give us the highest accuracy which is 99.45%. Hence, we can derive that the third model gives an accurate prediction.